refactor(operator): extract save-status helpers to slide_save module (#313)
slide_list.rs grew to 1026 lines after the per-slide Memo + Group-blur
sequencing fix, tripping the 1000-line CI hard cap. Move the save-status
state machine (SAVE_TOKEN, SaveStatusMap, start/finish helpers,
save_with_status) into a sibling slide_save.rs (92 lines). slide_list.rs
drops to 954 lines — under the hard cap, still over the 800 warn line
which was already the case before this PR.
