+++ crates/presenter-ui/src/components/slide_list.rs
@@ -1,7 +1,5 @@
 use std::collections::HashMap;
-use std::sync::atomic::{AtomicU64, Ordering};
 
-use gloo_timers::future::TimeoutFuture;
 use leptos::prelude::*;
 use presenter_core::{resolve_sequence, ResolvedSlide};
 use wasm_bindgen::JsCast;
@@ -16,11 +14,9 @@ use super::slide_list_utils::{
     apply_focused_class, field_has_warning, format_multiline, reorder_slide_ids,
     slide_has_any_warning,
 };
-
-/// Monotonic token used to invalidate stale fade timers (#313).
-/// Each save increments this and stamps the per-slide save_status entry;
-/// the 2s fade timer only removes the entry if the token still matches.
-static SAVE_TOKEN: AtomicU64 = AtomicU64::new(0);
+use super::slide_save::{
+    finish_save_status_err, finish_save_status_ok, save_with_status, start_save_status,
+};
 
 /// Get a textarea/input value from the DOM by slide_id and field name.
 fn get_field_value(doc: &web_sys::Document, slide_id: &str, field: &str) -> String {
@@ -132,74 +128,6 @@ fn save_all_fields_from_dom(
         group,
         op.save_status,
     );
-}
-
-/// Status-map signal alias used by the save-status helpers (#313).
-type SaveStatusMap = RwSignal<std::collections::HashMap<String, (SaveStatus, u64)>>;
-
-/// Mark a slide as `Saving` and return its monotonic token. Callers pass the
-/// token to `finish_save_status_ok` / `finish_save_status_err` to ensure a
-/// stale completion can't overwrite a newer save's entry (#313).
-fn start_save_status(slide_id: &str, save_status: SaveStatusMap) -> u64 {
-    let token = SAVE_TOKEN.fetch_add(1, Ordering::Relaxed) + 1;
-    let key = slide_id.to_string();
-    save_status.update(|map| {
-        map.insert(key, (SaveStatus::Saving, token));
-    });
-    token
-}
-
-/// Mark a slide save as `Saved` and schedule the 2s fade-removal. The fade
-/// only clears the entry when the token still matches — guards against an
-/// older fade timer clearing a newer save's badge.
-async fn finish_save_status_ok(slide_id: String, save_status: SaveStatusMap, token: u64) {
-    let key_for_saved = slide_id.clone();
-    save_status.update(|map| {
-        map.insert(key_for_saved, (SaveStatus::Saved, token));
-    });
-    TimeoutFuture::new(2_000).await;
-    save_status.update(|map| {
-        if map.get(&slide_id).map(|(_, t)| *t) == Some(token) {
-            map.remove(&slide_id);
-        }
-    });
-}
-
-/// Mark a slide save as `Failed` (sticky until the next save attempt).
-fn finish_save_status_err(slide_id: String, save_status: SaveStatusMap, token: u64) {
-    save_status.update(|map| {
-        map.insert(slide_id, (SaveStatus::Failed, token));
-    });
-}
-
-/// Fire-and-forget save with indicator wiring. Used by the three textarea
-/// blur handlers where no refetch is needed.
-fn save_with_status(
-    pres_id: String,
-    slide_id: String,
-    main: String,
-    translation: String,
-    stage: String,
-    group: Option<String>,
-    save_status: SaveStatusMap,
-) {
-    let token = start_save_status(&slide_id, save_status);
-    leptos::task::spawn_local(async move {
-        let result = api::presentations::update_slide_with_group(
-            &pres_id,
-            &slide_id,
-            &main,
-            &translation,
-            &stage,
-            group,
-        )
-        .await;
-
-        match result {
-            Ok(_) => finish_save_status_ok(slide_id, save_status, token).await,
-            Err(_) => finish_save_status_err(slide_id, save_status, token),
-        }
-    });
 }
 
 /// Capture current selection range from a textarea event.
